+++ src/easylink/steps/splink/splink_evaluating_pairs.py
@@ -33,6 +33,7 @@ for block_dir in blocks_dir.iterdir():
             comparisons.append(cl.DateOfBirthComparison(column))
         else:
             raise ValueError(f"Unknown comparison method {method}")
+    # TODO: check both datasets contain all the columns
 
     # Create the Splink linker in dedupe mode
     settings = SettingsCreator(